edit changes an project
