+++ .env.dev
@@ -6,4 +6,4 @@ PASSWORD=1111
 USER=accounter_web
 HOST=db
 PORT=5432
-NAME=accounter_postgres_db
+NAME=accounter_db
