+++ accounter/accounter/settings.py
@@ -24,8 +24,11 @@ INSTALLED_APPS = [
         'django.contrib.messages',
         'django.contrib.staticfiles'],
     *[
+        'rest_framework',],
+    *[
         'user',
-        'incomes']
+        'incomes',
+        'expenses']
 ]
 
 MIDDLEWARE = [
