+++ accounter/expenses/admin.py
@@ -1,3 +1,9 @@
 from django.contrib import admin
+from .models import ExpensesNote
 
-# Register your models here.
+class ExpensesNoteAdmin(admin.ModelAdmin):
+    list_display = ('id','user','score','where_from')
+    list_display_links = ('id','user','score','where_from')
+
+
+admin.site.register(ExpensesNote,ExpensesNoteAdmin)
