+++ accounter/incomes/admin.py
@@ -1,10 +1,9 @@
 from django.contrib import admin
-from .models import OneNote,Photo
+from .models import IncomesNote
 
-class IncomeNote(admin.ModelAdmin):
+class IncomesNoteAdmin(admin.ModelAdmin):
     list_display = ('id','user','score','where_from')
     list_display_links = ('id','user','score','where_from')
 
 
-admin.site.register(OneNote,IncomeNote)
-admin.site.register(Photo)
+admin.site.register(IncomesNote,IncomesNoteAdmin)
